# doc/api.yaml
openapi: 3.0.0
info:
    title: Home Library Service
    description: Home music library service
    version: 1.0.0
servers:
    - url: /
components:
    schemas:
...
        Artist:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: 'Freddie Mercury'
                grammy:
                    type: boolean
                    example: false
            required:
                - id
                - name
        Album:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: 'Innuendo'
                year:
                    type: integer
                    example: 1991
                artistId:
                    type: string
                    format: uuid
                    nullable: true
            required:
                - id
                - name
        Track:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: The Show Must Go On
                artistId:
                    type: string
                    format: uuid
                    nullable: true
                albumId:
                    type: string
                    format: uuid
                    nullable: true
                duration:
                    type: integer
                    description: In seconds
                    example: 262
            required:
                - id
                - name
                - duration
        Favorites:
            type: object
            properties:
                artists:
                    type: array
                    items:
                        $ref: '#/components/schemas/Artist'
                albums:
                    type: array
                    items:
                        $ref: '#/components/schemas/Album'
                tracks:
                    type: array
                    items:
                        $ref: '#/components/schemas/Track'
...
    responses:
        UnauthorizedError:
            description: Access token is missing or invalid
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
security:
    - bearerAuth: []
paths:
...
    /track:
        get:
            tags:
                - Track
            summary: Get tracks list
            description: Gets all library tracks list
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Track'
                401:
                    $ref: '#/components/responses/UnauthorizedError'
        post:
            tags:
                - Track
            summary: Add new track
            description: Add new track information
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                artistId:
                                    type: string
                                    format: uuid
                                albumId:
                                    type: string
                                    format: uuid
                                duration:
                                    type: integer
                                    description: In seconds
                            required:
                                - name
                                - duration
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Track'
                400:
                    description: Bad request. body does not contain required fields
                401:
                    $ref: '#/components/responses/UnauthorizedError'
    /track/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Track
            summary: Get single track by id
            description: Gets single track by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Track'
                400:
                    description: Bad request. trackId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
        put:
            tags:
                - Track
            summary: Update track information
            description: Update library track information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                name:
                                    type: string
                                    example: Bohemian Rhapsody
                                artistId:
                                    type: string
                                    format: uuid
                                    nullable: true
                                duration:
                                    type: integer
                                    description: In seconds
                                    example: 355
                                albumId:
                                    type: string
                                    format: uuid
                                    nullable: true
                            required:
                                - name
                                - duration
            responses:
                200:
                    description: The track has been updated.
                    content:
                        application/json:
                            schema:
                                type: object
                                title: example
                                properties:
                                    id:
                                        type: string
                                        format: uuid
                                    name:
                                        type: string
                                        example: Bohemian Rhapsody
                                    artistId:
                                        type: string
                                        format: uuid
                                        nullable: true
                                    duration:
                                        type: integer
                                        description: In seconds
                                        example: 355
                                    albumId:
                                        type: string
                                        format: uuid
                                        nullable: true
                400:
                    description: Bad request. trackId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
        delete:
            tags:
                - Track
            summary: Delete track
            description: Delete track from library
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. trackId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
...
    /album/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Album
            summary: Get single album by id
            description: Gets single album by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Album'
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
        put:
            tags:
                - Album
            summary: Update album information
            description: Update library album information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                name:
                                    type: string
                                year:
                                    type: integer
                                artistId:
                                    type: string
                                    format: uuid
                                    nullable: true
                            required:
                                - name
                                - year
            responses:
                200:
                    description: The album has been updated.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Album'
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
        delete:
            tags:
                - Album
            summary: Delete album
            description: Delete album from library
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
...
    /artist/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Artist
            summary: Get single artist by id
            description: Get single artist by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artist'
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
        put:
            tags:
                - Artist
            summary: Update artist information
            description: Update artist information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                grammy:
                                    type: boolean
                            required:
                                - name
                                - grammy
            responses:
                200:
                    description: The artist has been updated.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artist'
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
        delete:
            tags:
                - Artist
            summary: Delete artist
            description: Delete artist from library
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
    /favs:
        get:
            tags:
                - Favorites
            summary: Get all favorites
            description: Gets all favorites movies, tracks and books
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Favorites'
                401:
                    $ref: '#/components/responses/UnauthorizedError'
    /favs/track/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add track to the favorites
            description: Add track to the favorites
            responses:
                201:
                    description: Added successfully
                400:
                    description: Bad. trackId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Track with id doesn't exist.
        delete:
            tags:
                - Favorites
            summary: Delete track from favorites
            description: Delete track from favorites
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad. trackId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
    /favs/album/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add album to the favorites
            description: Add album to the favorites
            responses:
                201:
                    description: Added successfully
                400:
                    description: Bad. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Album with id doesn't exist.
        delete:
            tags:
                - Favorites
            summary: Delete album from favorites
            description: Delete album from favorites
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
    /favs/artist/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add artist to the favorites
            description: Add artist to the favorites
            responses:
                201:
                    description: Added successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Artist with id doesn't exist.
        delete:
            tags:
                - Favorites
            summary: Delete artist from favorites
            description: Delete artist from favorites
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.